+++ src/lib/libast/tm/tvgettime.c
@@ -2,7 +2,7 @@
 *                                                                      *
 *               This software is part of the ast package               *
 *          Copyright (c) 1985-2011 AT&T Intellectual Property          *
-*          Copyright (c) 2020-2023 Contributors to ksh 93u+m           *
+*          Copyright (c) 2020-2026 Contributors to ksh 93u+m           *
 *                      and is licensed under the                       *
 *                 Eclipse Public License, Version 2.0                  *
 *                                                                      *
@@ -14,6 +14,7 @@
 *                  David Korn <dgk@research.att.com>                   *
 *                   Phong Vo <kpv@research.att.com>                    *
 *                  Martijn Dekker <martijn@inlv.org>                   *
+*            Johnothan King <johnothanking@protonmail.com>             *
 *                                                                      *
 ***********************************************************************/
 
@@ -32,7 +33,7 @@ tvgettime(Tv_t* tv)
 
 	clock_gettime(CLOCK_REALTIME, &s);
 	tv->tv_sec = s.tv_sec;
-	tv->tv_nsec = s.tv_nsec;
+	tv->tv_nsec = (uint32_t)s.tv_nsec;
 
 #else
